+++ adapters/connectkit-connector/CHANGELOG.md
@@ -1,5 +1,52 @@
 # @blocto/connectkit-connector
 
+## 0.1.1-beta.5
+
+### Patch Changes
+
+- 39754e5: Fix disconnect all evm chains when disconnect
+- Updated dependencies [39754e5]
+  - @blocto/sdk@0.6.0-beta.5
+
+## 0.1.1-beta.4
+
+### Patch Changes
+
+- 6acf18d: Fix switch chain is connected logic
+- Updated dependencies [6acf18d]
+  - @blocto/sdk@0.6.0-beta.4
+
+## 0.1.1-beta.3
+
+### Patch Changes
+
+- 84d6727: Fix emit disconnect event when switch chain approved but failed
+- Updated dependencies [84d6727]
+  - @blocto/sdk@0.6.0-beta.3
+
+## 0.1.1-beta.2
+
+### Patch Changes
+
+- a0cc54c: Fix switch chain login logic
+- Updated dependencies [a0cc54c]
+  - @blocto/sdk@0.6.0-beta.2
+
+## 0.1.1-beta.1
+
+### Patch Changes
+
+- 0d5bfd5: Fix wrong switch chain url
+- Updated dependencies [0d5bfd5]
+  - @blocto/sdk@0.6.0-beta.1
+
+## 0.1.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [e0e8fb7]
+  - @blocto/sdk@0.6.0-beta.0
+
 ## 0.1.0
 
 ### Minor Changes